+++ app/src/main/java/com/example/movie/view/LoginActivity.kt
@@ -1,6 +1,7 @@
 package com.example.movie.view
 
 
+import android.content.Context
 import android.content.Intent
 import android.content.SharedPreferences
 import android.os.Bundle
@@ -8,6 +9,8 @@ import android.widget.Button
 import android.widget.EditText
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProvider
 import com.example.movie.BuildConfig
 import com.example.movie.R
 import com.example.movie.api.RequestToken
@@ -16,6 +19,9 @@ import com.example.movie.api.Session
 import com.example.movie.model.MyAccount
 import com.example.movie.model.Singleton
 import com.example.movie.model.User
+import com.example.movie.view_model.LikeListViewModel
+import com.example.movie.view_model.LoginViewModel
+import com.example.movie.view_model.ViewModelProviderFactory
 import com.google.gson.Gson
 import com.google.gson.JsonObject
 import com.google.gson.reflect.TypeToken
@@ -24,26 +30,28 @@ import java.lang.Exception
 import java.lang.reflect.Type
 import kotlin.coroutines.CoroutineContext
 
-class LoginActivity : AppCompatActivity(), CoroutineScope by MainScope() {
+class LoginActivity : AppCompatActivity() {
     private lateinit var email: EditText
     private lateinit var password: EditText
     private lateinit var login: Button
     private lateinit var register: Button
     private lateinit var preferences: SharedPreferences
-    private lateinit var requestToken: String
-    private lateinit var newRequestToken: String
+    private lateinit var json1: String
     private lateinit var emailValue: String
     private lateinit var passwordValue: String
-    private val job = Job()
-
-    override val coroutineContext: CoroutineContext
-        get() = Dispatchers.Main + job
+    private lateinit var loginViewModel: LoginViewModel
+    private var sessionId: String? = null
+    private var accountId: Int? = 0
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_login)
+        val viewModelProviderFactory = ViewModelProviderFactory(context = this)
+        loginViewModel =
+            ViewModelProvider(this, viewModelProviderFactory).get(LoginViewModel::class.java)
         bindView()
         stayLogged()
+
 
         login.setOnClickListener {
             loginCoroutine()
@@ -53,74 +61,37 @@ class LoginActivity : AppCompatActivity(), CoroutineScope by MainScope() {
             val intent = Intent(this, RegistrationActivity::class.java)
             startActivity(intent)
         }
+        loginViewModel.liveDataLogin.observe(this, Observer { result ->
+            json1 = result[0]
+            sessionId = result[1]
+            accountId = Integer.parseInt(result[2])
+
+        })
     }
 
-    override fun onDestroy() {
-        super.onDestroy()
-        job.cancel()
-    }
 
     private fun loginCoroutine() {
         emailValue = email.text.toString()
         passwordValue = password.text.toString()
         if (emailValue != "" && passwordValue != "") {
-            launch {
-                val response = RetrofitService.getPostApi()
-                    .getRequestTokenCorountine(BuildConfig.THE_MOVIE_DB_API_TOKEN)
-                if (response.isSuccessful) {
-                    requestToken = response.body()?.requestToken!!
-                    val body = JsonObject().apply {
-                        addProperty("username", emailValue)
-                        addProperty("password", passwordValue)
-                        addProperty("request_token", requestToken)
-                    }
-
-                    val responseLogin = RetrofitService.getPostApi()
-                        .loginCoroutune(BuildConfig.THE_MOVIE_DB_API_TOKEN, body)
-
-                    if (responseLogin.isSuccessful) {
-                        var gson = Gson()
-                        var newRequesttoken = gson.fromJson(
-                            responseLogin.body(),
-                            RequestToken::class.java
+            loginViewModel.login(emailValue, passwordValue)
+            try {
+                var MySingleton = sessionId?.let {
+                    accountId?.let { it1 ->
+                        Singleton.create(
+                            emailValue, it,
+                            it1
                         )
-                        newRequestToken = newRequesttoken.requestToken
-                        val responseSession = RetrofitService.getPostApi()
-                            .getSessionCoroutine(BuildConfig.THE_MOVIE_DB_API_TOKEN, body)
-
-                        if (responseSession.isSuccessful) {
-                            var gson = Gson()
-                            var newSession =
-                                gson.fromJson(
-                                    responseSession.body(),
-                                    Session::class.java
-                                )
-                            val sessionId = newSession.sessionId
-                            val response = RetrofitService.getPostApi()
-                                .getAccountCoroutine(BuildConfig.THE_MOVIE_DB_API_TOKEN, sessionId)
-                            if (response.isSuccessful) {
-                                var gson = Gson()
-                                var newIdAcc =
-                                    gson.fromJson(response.body(), MyAccount::class.java)
-                                var idAcc = newIdAcc.id
-                                var user = User(emailValue, sessionId, idAcc)
-                                var MySingleton = Singleton.create(emailValue, sessionId, idAcc)
-                                val json1 = gson.toJson(user)
-                                preferences = this@LoginActivity.getSharedPreferences("Username", 0)
-                                preferences.edit().putString("user", json1).commit()
-                                val intent = Intent(this@LoginActivity, MainActivity::class.java)
-                                startActivity(intent)
-                            } else {
-                                Toast.makeText(
-                                    this@LoginActivity,
-                                    "Fail",
-                                    Toast.LENGTH_SHORT
-                                ).show()
-                            }
-                        }
                     }
                 }
+                preferences = this@LoginActivity.getSharedPreferences("Username", 0)
+                preferences.edit().putString("user", json1).commit()
+                val intent = Intent(this@LoginActivity, MainActivity::class.java)
+                startActivity(intent)
+            } catch (e: Exception) {
+
             }
+
         } else {
             Toast.makeText(this@LoginActivity, "Empty email or password", Toast.LENGTH_LONG)
                 .show()
